0422yujeong

+++ public/scripts/search.js
@@ -16,6 +16,9 @@ function ready() {
     */
 }
 function deleteRecipe(event) {
+    if(!confirm('Are you sure you want to delete this recipe?')){
+        return false;
+    }
     const db = getFirestore();
     db.collection('Recipes').doc(event).delete();
     var buttonClicked = event.target
